+++ fourier/Fourier_DCT4.c
@@ -10,8 +10,6 @@
 # include <xmmintrin.h>
 #endif
 /**************************************/
-#include <stddef.h>
-/**************************************/
 #include "Fourier.h"
 /**************************************/
 
@@ -116,8 +114,8 @@ static void DCT4_8(float *x) {
 
 /**************************************/
 
-void Fourier_DCT4(float *Buf, float *Tmp, size_t N) {
-	size_t i;
+void Fourier_DCT4(float *Buf, float *Tmp, int N) {
+	int i;
 
 	//! Stop condition
 	if(N == 8) {
